feat: Enhance user management and password reset functionality
This commit introduces several key enhancements to the user management system:

- **User Schema Update**: Adds `passwordResetToken` and `passwordResetTokenExpiry` properties to the User schema in `user.model.js`. These fields are essential for implementing password reset functionality, allowing users to securely reset their passwords.

- **Utility Functions Migration**: Moves the definitions of `generateAccessAndRefreshTokens` and `validateEmail` functions from `user.controller.js` to a newly created file `user.utils.js`. This change organizes utility functions in a dedicated file, improving code maintainability and readability.

- **Password Reset Utilities**: Adds two new utility functions, `generatePasswordResetToken` and `sendPasswordResetEmail`, in `user.utils.js`. These functions facilitate the password reset process by generating a unique token for password reset requests and sending the reset link to the user's email.

- **Password Reset Controllers**: Introduces two new controller functions, `forgetPasswordEmail` and `resetPassword`, in `user.controller.js`. These functions handle the process of initiating a password reset request and resetting the user's password, providing a secure and user-friendly way to recover lost passwords.

These updates significantly enhance the application's security and user experience by implementing a robust password reset mechanism, ensuring users can securely recover their accounts if they forget their passwords.
