+++ src/routes/user.routes.js
@@ -4,13 +4,6 @@ import {
     loginUser,
     logoutUser,
     updateAccessToken,
-    updateUserDetails,
-    updateUserPassword,
-    updateAvatar,
-    updateCoverImage,
-    getCurrentUser,
-    forgetPasswordEmail,
-    resetPassword,
 } from "../controllers/user.controller.js";
 import { upload } from "../middlewares/multer.middleware.js";
 import { verifyJWT } from "../middlewares/auth.middleware.js";
@@ -34,6 +27,5 @@ router.route("/registration").post(
 router.route("/login").post(loginUser);
 router.route("/logout").post(verifyJWT, logoutUser);
 router.route("/access-token").post(updateAccessToken);
-
 
 export default router;
